requirements.txt
annotated-types==0.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.5.1
typing_extensions==4.12.2
urllib3==2.2.1
wheel==0.43.0
